Implement transactions for accumulating batchPlay calls
descriptor.beginTransaction initiates a transaction and returns an object, which is supplied
to different batchPlay calls as the `transaction` key in play options.

At the end of all the action calls, the `descriptor.endTransaction` promise will resolve to play
all the batchPlay objects accumulated underneath that transaction.

Use with caution, all the actions/batchPlay you're calling should not be dependent on each other,
or results from Photoshop's models changing.

- Implemented by @iwehrman, rebased/described by @volfied